+++ python/polars_online/__init__.py
@@ -46,7 +46,6 @@ from polars_online._bank import ModelBank
 from polars_online._expr import InMemoryExpressionWarning, online
 from polars_online._frame import fit_predict, predict, unnest
 from polars_online._polars_online import native_version, schema_version, thread_pool_size
-from polars_online._runner import run
 
 __version__ = "0.6.0"
 
@@ -62,7 +61,6 @@ __all__ = [
     "online",
     "predict",
     "prep",
-    "run",
     "schema_version",
     "sim",
     "spec",